Harden Geo Ads Pro plugin

+++ uninstall.php
@@ -12,13 +12,23 @@ if (!defined('WP_UNINSTALL_PLUGIN')) {
     exit;
 }
 
-// Delete options from database
+$delete_data = (int) get_option('gap_delete_data_on_uninstall', 0);
+
+if (!$delete_data) {
+    return;
+}
+
 delete_option('gap_enable_local_mode');
 delete_option('gap_default_region');
 delete_option('gap_rotation_mode');
+delete_option('gap_rotation_state');
 delete_option('gap_abtest_auto');
+delete_option('gap_delete_data_on_uninstall');
+delete_option('gap_version');
+delete_option('gap_schema_version');
+delete_option('gap_upgraded_at');
+delete_option('gap_upgrade_required');
 
-// Delete uploads folder and all files inside
 $upload = wp_upload_dir();
 $dir = trailingslashit($upload['basedir']) . 'geo-ads-pro';
